Update WiFi Basic Example
Changed the int before prevWindTelemetry to an unsigned long.

+++ examples/iotQi-WiFi-basic/iotQi-WiFi-basic.ino
@@ -39,7 +39,7 @@ void setup() {
 
 
 /* loop() ------------------------------------------------------------------------------------------------------------- */
-int prevWindTelemetry = 0;
+unsigned long prevWindTelemetry = 0;
 
 void loop() 
 {
